requirements.txt
joblib>=1.3
matplotlib>=3.6
numpy>=1.23
pandas>=1.5
python-dotenv>=1.0
requests>=2.31
scikit-learn>=1.2
